Modified height and width
