--- packages/web-client/src/components/backstory/Backstory.module.css
@@ -1,22 +1,0 @@
-
-.story {
-    height: 70vh;
-    width: 65vw;
-    background: linear-gradient(180deg, rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6), rgba(0, 0, 0, 0.6), transparent);
-    padding: 20px 50px;
-    margin-left: auto;
-    margin-right: auto;
-    margin-top: auto;
-}
-
-.text {
-    font-family: 'Alexandria', sans-serif;
-    text-align: justify;
-    font-weight: 350;
-    font-size: 1.3em;
-    line-height: 1.5;
-}
-
-.city {
-    font-size: 2.7em;
-}
