+++ packages/web-client/src/pages/Onboarding.tsx
@@ -3,7 +3,7 @@ import About from '../components/About';
 import Footer from '../components/footer/Footer';
 import Roadmap from '../components/roadmap/Roadmap';
 import Team from '../components/team/Team';
-import Backstory from '../components/backstory/Backstory';
+import Backstory from '../components/Backstory';
 
 const Onboarding = () => {
     return(
